Change interface to be Iterable-based
This changes the interface to be Iterable-based.  It makes a heroic
assumption that we will always access the entries in strict order, so
the iterator's `.next()` method will always return exactly the entry
we're interested in.

This assumption holds for computing the currentRoot() but not for any of
the proofs, so those tests are now broken.

This change is so that we can consume an Iterator which does batched
querying of the database rather than running a separate database query
for each entry.